+++ lib/financial_tracker_app/models/transaction.dart
@@ -2,7 +2,7 @@ class TransactionModel {
   final int? id;
   final int categoryId;
   final String categoryName;
-  final double amount;
+  final int amount;
   final bool isIncome;
   final String hijriMonth; // e.g. "1447-10"
   final int createdAt; // epoch millis
@@ -31,7 +31,7 @@ class TransactionModel {
     id: m['id'] as int?,
     categoryId: m['category_id'] as int,
     categoryName: m['category_name'] as String,
-    amount: (m['amount'] as num).toDouble(),
+    amount: (m['amount'] as num).toInt(),
     isIncome: (m['is_income'] as int) == 1,
     hijriMonth: m['hijri_month'] as String,
     createdAt: m['created_at'] as int,
